first parallel implementation

+++ enc/include/chacha.hpp
@@ -12,8 +12,8 @@ class ChaCha20 {
         ChaCha20(const uint8_t key[32], const uint8_t nonce[12]);
         void setCounter(uint32_t counter);
         void block_quarter_round(uint8_t result[64], uint32_t counter);
-        uint8_t *encrypt(uint8_t *input, int len);
-        uint8_t *decrypt(uint8_t *input, int len);
+        uint8_t *encrypt(uint8_t *input, long len);
+        uint8_t *decrypt(uint8_t *input, long len);
     private:
         void _encrypt_single_block(uint8_t *input, uint8_t *result, uint32_t counter);
 };
